Several changes and improvements to the command manager.
Multiple optional arguments can be used in a command, and the command
manager will try to automatically determine which optional arguments are
being supplied by the sender.

If the name of the command argument type and the name of the command
argument on the same, the type will automatically be hidden.

For default arguments, you can now specify a default value by adding a
string which will be parsed into the value in parenthesis after the
argument. Example:

commandname string:string?(default value here) {