Removed on Roll feature
Option to enable removeOnRoll, which consist on removing the rolled object when calling roll() so is neved rolled again. I came to this problem in one of my projects, where i wanted a weightedRandom object from a collection but to never appear again. This could be also implemented from the api-user storing the rolled items and if the roll() output is contained, roll again, but i found this to be a better solution since my collection is not really that big.

To sum it up is not a big feature but i wanted to collaborate on the project since i found it really useful on my projects, thaks for all man!