Remove lore functionality (#49)
* Remove lore functionality

Added four new methods with documentation. In ItemUtils I added removing lore from an itemstack with either a specific lore line or index of the line. This will do nothing if the lore line doesnt exist or index is out of bounds. I implemented this into ItemBuilder but I don't know how useful that'll be.

* Removed useless line, added out of bounds throwing

* removed cloning of list 😅

* got rid of white space